A bit of bug fixing.
It appears that the install, deactive, and unistall methods were not
being called, or properly hooked. According to the wordpress
documentation those hooks need to point to the plugin's main file. So I
moved them into there.

This resolved the issue with various menus not appearing in the
Dashboard.

I also changed the "Order Profiles" tab so that it just displays the
name of the profile, there seems to be some issue with the
edit_post_link function, that is causing it to output just an empty
string, so I commented it out for the moment.

Also just some adjustment to the spacing in various files as I was
reading them.