Changed the behavious of the Refresh Profiles dialog
There is a dialog that appears at the top of the page whenever you
modify the profile page or list layouts. I've changed it so that the
message will not go away if you refresh the page, and will consistently
tell you, which layouts have changed and provide the update button.

I used the wordpress Transients API, currently the alert is set to
expire in 3 days if the user chooses not to press the update button.

Currently the button doesn't do anything because the necessary function
hasn't been created (or moved from the old code).