Fixed issues with the form and views.
Fixed a variety of issues with the front end views, and the profile
editing form. All the field types should not edit and display properly.

The layout of the front end page hasn't been fixed yet.

Additionally the ajax updating of posts seems to work now, but the
progress bar is untested.